refactor: reorganize routes and improve configuration management
This commit reorganizes the library's structure to improve maintainability
and user experience:

- Rename /summary routes to /user for better semantic clarity
- Add centralized configuration with O2P_REDIRECT_ANON and O2P_REDIRECT_USER
- Move page handlers to a dedicated pages module with proper organization
- Simplify demo integration by using library-provided redirects
- Fix formatting and add trace logging in security checks
- Hide user ID from summary page UI for better security
- Consolidate templates and remove redundant code

This refactoring makes the library more modular, easier to maintain, and
provides better defaults for integrators while maintaining all existing
functionality.
