--- demo-integration/templates/index_user.j2
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Index Page</title>
-    <style>
-        {# .user-info {
-            background: #f5f5f5;
-        }
-        .user-info dt {
-            font-weight: bold;
-            margin-top: 0.5rem;
-        }
-        .user-info dd {
-            margin-top: 0.5rem;
-        }
-        .user-item {
-            margin-top: 0.5rem;
-        } #}
-        .user-passkeys {
-            margin-top: 0.5rem;
-        }
-        .user-accounts {
-            margin-top: 0.5rem;
-        }
-    </style>
-</head>
-
-<body>
-    <div>
-        <h1>{{message}}</h1>
-        <p><a href="/summary" style="color: #4285f4; text-decoration: none; font-weight: bold;">View Your Complete Account Summary</a></p>
-    </div>
-
-    {# <div class="user-info">
-        <dl>
-            <dt>Your User Information</dt>
-            <dd>
-                <div class="user-item"><strong>User ID:</strong> {{ user.id }}</div>
-                <div class="user-item"><strong>Name:</strong> {{ user.name }}</div>
-                <div class="user-item"><strong>Created:</strong> {{ user.created_at }}</div>
-                <div class="user-item"><strong>Updated:</strong> {{ user.updated_at }}</div>
-            </dd>
-        </dl>
-    </div> #}
-    <div class="user-passkeys">
-        <button onclick="showRegistrationModal('add_to_existing_user')">Add New Passkey</button>
-    </div>
-    <div class="user-accounts">
-        <button onclick="oauth2.openPopup()">Add New OAuth2 Account</button>
-    </div>
-
-    <div style="display: flex; gap: 10px; margin-top: 10px;">
-        <button onclick="Logout()">Logout</button>
-        <script>
-            function Logout() {
-                window.location.href = "{{o2p_route_prefix}}/oauth2/logout";
-            }
-        </script>
-    </div>
-
-    <script src="{{o2p_route_prefix}}/oauth2/oauth2.js"></script>
-    <script src="{{o2p_route_prefix}}/passkey/passkey.js"></script>
-    <script>
-        const oauth2 = initOAuth2Popup();
-        const O2P_ROUTE_PREFIX = '{{o2p_route_prefix}}';
-    </script>
-
-</body>
-</html>
