+++ oauth2_passkey_axum/src/pages/mod.rs
@@ -1,0 +1,5 @@
+mod login;
+mod router;
+mod user;
+
+pub use router::router as user_router;
