--- oauth2_passkey_axum/src/summary/mod.rs
@@ -1,4 +1,0 @@
-mod handlers;
-mod templates;
-
-pub use handlers::router;
